+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.9.x
   - Fix events not being processed in a timely manner
   - Mark plugin as non-reloadable for https://github.com/elastic/logstash/pull/6499
+  - Add ECS compatibility support v1 and v8
 
 ## 0.9.1
   - Initial version of logstash-input-fifo