+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.9.x
+## 0.9.6
   - Fix events not being processed in a timely manner
   - Mark plugin as non-reloadable for https://github.com/elastic/logstash/pull/6499
   - Add ECS compatibility support v1 and v8